add Author model
