+++ backend/apps/books/models.py
@@ -1,5 +1,90 @@
 from django.core.validators import RegexValidator
 from django.db import models
+from django.db.models import Avg
+from django.db.models.query import QuerySet
+
+# from django.urls import reverse
+from django.utils.translation import gettext_lazy as _
+from django_extensions.db.fields import (
+    AutoSlugField,
+    CreationDateTimeField,
+    ModificationDateTimeField,
+)
+
+# TODO in progress
+# temporary here, move to books/models if available?
+
+# class AuthorManager(models.Manager):
+#     """Queryset manager for Author model."""
+
+#     # def get_queryset(self) -> QuerySet:
+#     #     """Build-in queryset"""
+#     #     return super().get_queryset()
+
+#     def genres(self) -> QuerySet:
+#         """Returns genres that author's books belong to."""
+#         # TODO query to be checked, not sure it works this way
+#         return self.books.genres.all().distinct()
+
+#     def scoring(self) -> float:
+#         """Returns average of author's books' scoring."""
+#         # TODO query to be checked, not sure it works this way
+#         return self.books.reviews.all().aggregate(Avg("score"))
+
+
+class Author(models.Model):
+    """Model for books' authors."""
+
+    class Meta:
+        verbose_name = _("author")
+        verbose_name_plural = _("authors")
+        ordering = ("first_name",)
+
+    first_name = models.CharField(max_length=64, verbose_name=_("first name"))
+    middle_name = models.CharField(
+        max_length=64, null=True, blank=True, verbose_name=_("middle name")
+    )
+    last_name = models.CharField(max_length=64, verbose_name=_("last name"))
+    full_name = models.CharField(max_length=255, verbose_name=_("full name"))
+    slug = AutoSlugField(populate_from=["first_name", "last_name"])
+
+    birth_date = models.DateField(
+        null=True, blank=True, verbose_name=_("date of birth")
+    )
+    death_date = models.DateField(
+        null=True, blank=True, verbose_name=_("date of death")
+    )
+    description = models.TextField(null=True, blank=True, verbose_name=_("description"))
+    allow_highlights = models.BooleanField(
+        null=False, blank=False, default=True, verbose_name=_("allow highlights")
+    )
+    created = CreationDateTimeField()
+    modified = ModificationDateTimeField()
+
+    # objects = AuthorManager()
+
+    # from apps.generic.modles import Highlight, Image
+    # images = GenericRelation(Image, related_query_name="authors")
+    # highlights = GenericRelation(Highlight, related_query_name="authors")
+
+    def __str__(self) -> str:
+        return self.full_name
+
+    # trial
+    def make_full_name(self) -> str:
+        """Author's full name."""
+        if not self.mid_name:
+            return f"{self.first_name} {self.last_name}"
+        return f"{self.first_name} {self.mid_name} {self.last_name}"
+
+    def get_absolute_url(self) -> str:
+        # return reverse("authors-detail", args=[self.slug])
+        pass
+
+    # TODO check the save function.
+    def save(self, *args, **kwargs):
+        self.full_name = self.make_full_name()
+        return super().save(*args, **kwargs)
 
 
 class Genre(models.Model):
